requirements.txt
flake8==5.0.4
flake8-docstrings==1.6.0
pytest==7.1.3
python-dotenv==0.21.0
python-telegram-bot==13.14
requests==2.28.1
